Rewrite SBJsonTokeniser.
This version does not do the ugly hackery to keep the utf8 stream NUL-terminated. It has less duplication of logic since it does the parsing and decoding of strings and numbers in one pass now, rather than as two separate passes. I've not measured performance, but this should in theory be faster. (And even if it isn't it's worth it for the massive cleanup.)