Revert "Convert all library code to use Automatic Reference Counting (ARC)"
This reverts commit 74fa6866e478b7ad0ad80cf06f6c57335ae4a6e7.
It turns out that I was wrong; there's a run-time component involved with ARC, so if you convert to ARC you appear to not be able to compile for iOS4.3. (And it's much, much, slower compiled for 5.0 than for 4.3.)

As a work-around, if you want to copy the classes into your project you can just convert to ARC yourself. Or, you can link to the library.