Patch SBJsonStreamParser for integral values.
SBJsonStreamParser parses an integer into an NSDecimalNumber. With some
large values, such as Twitter tweet ids, this results in a rounding
error. This pull request makes a one line change to the parser to parse
into a longLongNumber instead of a decimal number. It also modifies the
sample app to use the Apple social media APIs and accounts. The problem
can be demonstrated by uncommenting the original line in the
SBJsonStreamParser file. Then the app will start logging tweets that
have incorrect values for tweet ids.