Rename classes, constants & enums to add major version number (4)
This is to allow SBJson4 to be used alongside previous versions of itself. This
can be useful now that SBJson is embedded by so many popular libraries.
Additionally, if you're already using it you may not want to replace all usages
in one go; this way you don't have to.