Recreate Xcode project & targets & change to a "more traditional" repo layout
Hopefully this will lower the barrier to contribution for people.

Upgrade to settings recommended by Xcode 8.1. As the test targets no
longer use a precompiled header, we have to import the XCTest header.

We also add TestData as a file reference, to avoid having so many
references to test data files in the Xcode project file. (Smaller
project file -> less merge conflicts.) We need to tweak tests slightly
to add the TestData directory to the path, but I prefer that to having
so many references.

Finally update the TravisCI build job to work with the two new targets.