Add guard for theoretical buffer overflow
I was wrong in the previous comment here. `snprintf` returns the
length *it would have written* if the buffer was big enough, which you
can use to detect overflow.

I can't actually see how an overflow can occur here in practice, but I
happened to notice this and thought I'd correct it for the record.